+++ src/Nox.Api/ApiRoute.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations.Schema;
 using Nox.Core.Components;
 using Nox.Core.Interfaces;
