+++ src/Nox.Core/Components/ServiceMessageBus.cs
@@ -1,9 +1,5 @@
-﻿
-using Nox.Core.Interfaces;
-using Nox.Core.Models;
+﻿namespace Nox.Core.Components;
 
-namespace Nox.Core.Components;
-
-public sealed class ServiceMessageBus : Models.MessageBusBase { }
+public sealed class ServiceMessageBus : ServiceMessageBusBase { }
