+++ src/Nox.Core/Components/ServiceMessageBusBase.cs
@@ -1,10 +1,8 @@
-using System.ComponentModel.DataAnnotations.Schema;
-using Nox.Core.Components;
 using Nox.Core.Interfaces;
 
-namespace Nox.Core.Models;
+namespace Nox.Core.Components;
 
-public class MessageBusBase : MetaBase, IServiceMessageBus
+public class ServiceMessageBusBase : MetaBase, IServiceMessageBus
 {
     public string Name { get; set; } = string.Empty;
     public string Provider { get; set; } = string.Empty;
