+++ src/Nox.Core/Models/MergeState.cs
@@ -1,5 +1,4 @@
 ﻿using Nox.Core.Components;
-using Nox.Core.Interfaces;
 
 namespace Nox.Core.Models
 {
