+++ src/Nox.Data/DynamicDbEntity.cs
@@ -1,6 +1,5 @@
 using System.Reflection;
 using System.Reflection.Emit;
-using Nox.Core.Components;
 using Nox.Core.Interfaces;
 
 namespace Nox.Data;
