+++ src/Nox.Etl/EtlExecutor.cs
@@ -6,8 +6,8 @@ using ETLBox.DataFlow.Connectors;
 using MassTransit;
 using MassTransit.Mediator;
 using Microsoft.Extensions.Logging;
-using Nox.Core.Components;
 using Nox.Core.Interfaces;
+using Nox.Core.Models;
 using Nox.Messaging;
 using Nox.Messaging.Enumerations;
 using Nox.Messaging.Events;
