+++ src/Nox.Etl/LoaderSchedule.cs
@@ -1,5 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
-using Nox.Core.Components;
+﻿using Nox.Core.Components;
 using Nox.Core.Interfaces;
 using Nox.Cron;
