+++ src/Nox.Messaging/Events/INoxEvent.cs
@@ -1,5 +1,3 @@
-using Nox.Core.Interfaces;
-
 namespace Nox.Messaging.Events;
 
 public interface INoxEvent
