+++ src/Nox.Microservice/Extensions/DynamicModelExtensions.cs
@@ -1,0 +1,11 @@
+using Nox.Core.Interfaces;
+
+namespace Nox.Microservice.Extensions;
+
+public static class DynamicModelExtensions
+{
+    public static void EnsureDatabaseCreated(this IMetaService metaService)
+    {
+        
+    }
+}
