+++ src/Nox.Microservice/Validation/LoaderSourceValidator.cs
@@ -1,6 +1,3 @@
-using FluentValidation;
-using Nox.Core.Models;
-
 namespace Nox.Microservice.Validation;
 
 public class LoaderSourceValidator : DatabaseValidator
