+++ src/Nox.Microservice/Validation/LoaderValidator.cs
@@ -1,7 +1,5 @@
 using FluentValidation;
-using Nox.Core.Components;
 using Nox.Core.Interfaces;
-using Nox.Core.Models;
 
 namespace Nox.Microservice.Validation;
