+++ src/Nox.Microservice/Validation/MetaServiceValidator.cs
@@ -1,5 +1,4 @@
 using FluentValidation;
-using Nox.Core.Models;
 
 namespace Nox.Microservice.Validation;
