+++ tests/Nox.Data.Tests/MySqlDatabaseProviderTests.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
-using Nox.Core.Components;
 using Nox.Core.Interfaces;
+using Nox.Core.Models;
 using Nox.TestFixtures;
 using NUnit.Framework;
