+++ tests/Nox.Generator.TestFixture/GeneratorTestFixture.cs
@@ -1,5 +1,4 @@
 using System.IO;
-using System.Reflection;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using NUnit.Framework;
