+++ tests/Nox.TestFixtures/EtlTestFixture.cs
@@ -1,9 +1,7 @@
 using System;
-using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Nox.Data;
-using Nox.Microservice;
 using Nox.Microservice.Extensions;
 using NUnit.Framework;
