+++ tests/Nox.TestFixtures/MessagingTestFixture.cs
@@ -1,10 +1,7 @@
 using System;
-using System.Reflection;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Nox.Data;
-using Nox.Messaging;
-using Nox.Microservice;
 using Nox.Microservice.Extensions;
 using NUnit.Framework;
