+++ tests/Nox.TestFixtures/NoxTestFixture.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Nox.Core.Interfaces;
-using Nox.Microservice;
 using Nox.Microservice.Extensions;
 using NUnit.Framework;
